+++ frontend/src/features/Result/Result.tsx
